Generates Windows deployable package
Adds .wxs file used to generate a .msi file with WiX Toolset.
Adds a CI configuration file that builds the .msi file and uploads it to
GitHub when running on git tags.

The second step requires a environment variable to be configured on the
Travis CI side with a personal access token with `public_repo` access.

The general deploy flow is as simple as:
 * `git tag 0.0.1`
 * `git push`

An msi file should be automatically created in GitHub project releases.